# old/api/openapi.yaml
openapi: 3.0.4
info:
  title: Depot catalog
  description: |
    # Информация о учебном процессе
    - Специальности
    - Группы
    - Студенты
    - Преподаватели
    - Дисциплины
    - Аудитории
  version: 1.0.0
servers:
  - url: http://localhost:18081
...
paths:
...
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system.
      description: Log into the system.
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: false
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: false
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        '400':
          description: Invalid username/password supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message